refactor: BucketListCard header, content, footer 카드로 분리
